From: Pythagoras Watson <py@llnl.gov> Date: Fri, 3 Oct 2008 15:52:42 -0700 To: Mark Grondona <mgrondona@llnl.gov>
One other thing, the perl-DBD-MySQL "Requires" in the sqlog.spec fails
on SLES, since the RPM is called perl-DBD-mysql.  I have attached a
patch that changes the various perl Requires to be based on the module
name rather than the RPM name.  Based off of my testing, it works for
all of Chaos 3.3, Chaos 4.0, and SLES 9.

FWIW, the new requires are mostly duplicates of what is auto-found by
RPM for Chaos 3 and 4 (which won't hurt anything), but the RPM on SLES
is old enough that that feature wasn't added yet.  The
perl(DBD::mysql) would still be necessary though, since it isn't
brought in via "use".

==Py