sqlog-db-util: Fix error in initial DB setup
When creating the slurm DB for the first time, the database handle
was being passed undefined to create_slurm_joblog_table_v2(), causing
the script to abort with runtime errors.

Switch the logic so that the DB connection is attempted *before* trying
to recreate the slurm joblog tables to avoid the error.