Do not trust SLURM NODECNT variable
When running the jobcomp/script, SLURM ostensibly sets the
NODECNT variable to number of nodes allocated to the completing
job. However, this variable has been found to be incorrect in
the case of NODE_FAIL jobs, because it is decremented by one
before the job completion script is run.

Therefore we have to always manually count the number of nodes
in the job.