sqlog-db-util: add --recalc-nodecnt
Some jobs in the sqlog database or SLURM joblog may have the
incorrect nodecount due a bug in how slurm sets the NODECNT
variable for jobcomp/script scripts. In order to fix the database
a new parameter is added to sqlog-db-util, --recalc-nodecnt, which
causes sqlog-db-util to recalculate the nodecount for every job
from the nodelist during a backfill operation. This allows one to
"fix" the database by doing something like:

  sqlog-db-util --backup=all joblog
  sqlog-db-util --drop=2
  sqlog-db-util -L --recalc-nodecount --backfill joblog