Support the RESIZING job state
When a job with --no-kill has a node that goes DOWN, slurm
puts the job into RESIZING, removes the down node and then
puts the job back into RUNNING. This generates a job completion
log with the state of RESIZING, which currently causes sqlog to
generate and error.

This small patch just does the bare minimum to handle the new
RESIZING state.